[BUGFIX] Long for UID

+++ src/main/java/org/rev317/min/accessors/SceneObjectTile.java
@@ -2,5 +2,5 @@ package org.rev317.min.accessors;
 
 public interface SceneObjectTile {
 
-    int getHash();
+    long getHash();
 }
